Separate common part out and align demo_config folder with Linux application.
