+++ examples/app_common/app_common.h
@@ -14,27 +14,27 @@
  * limitations under the License.
  */
 
-#ifndef DEMO_DATA_TYPES_H
-#define DEMO_DATA_TYPES_H
-
-#pragma once
-
-#ifdef __cplusplus
-extern "C" {
-#endif
+#ifndef APP_COMMON_H
+#define APP_COMMON_H
 
 #include <stdio.h>
 #include "sdp_controller.h"
 #include "signaling_controller.h"
 #include "peer_connection.h"
-#include "app_media_source.h"
 
 #define DEMO_SDP_BUFFER_MAX_LENGTH ( 10000 )
 #define DEMO_TRANSCEIVER_MEDIA_INDEX_VIDEO ( 0 )
 #define DEMO_TRANSCEIVER_MEDIA_INDEX_AUDIO ( 1 )
 #define REMOTE_ID_MAX_LENGTH    ( 256 )
 
-typedef struct DemoPeerConnectionSession
+struct AppMediaSourcesContext;
+typedef struct AppMediaSourcesContext AppMediaSourcesContext_t;
+
+typedef int32_t ( * InitTransceiverFunc_t )( void * pCtx,
+                                             TransceiverTrackKind_t trackKind,
+                                             Transceiver_t * pTranceiver );
+
+typedef struct AppSession
 {
     /* The remote client ID, representing the remote peer, from signaling message. */
     char remoteClientId[ REMOTE_ID_MAX_LENGTH ];
@@ -46,9 +46,12 @@ typedef struct DemoPeerConnectionSession
     /* Peer connection session. */
     PeerConnectionSession_t peerConnectionSession;
     Transceiver_t transceivers[ PEER_CONNECTION_TRANSCEIVER_MAX_COUNT ];
-} DemoPeerConnectionSession_t;
 
-typedef struct DemoContext
+    /* Initialized signaling controller. */
+    SignalingControllerContext_t * pSignalingControllerContext;
+} AppSession_t;
+
+typedef struct AppContext
 {
     /* Signaling controller. */
     SignalingControllerContext_t signalingControllerContext;
@@ -59,12 +62,16 @@ typedef struct DemoContext
     char sdpBuffer[ PEER_CONNECTION_SDP_DESCRIPTION_BUFFER_MAX_LENGTH ];
 
     /* Peer Connection. */
-    DemoPeerConnectionSession_t peerConnectionSessions[ AWS_MAX_VIEWER_NUM ];
-    AppMediaSourcesContext_t appMediaSourcesContext;
-} DemoContext_t;
+    AppSession_t appSessions[ AWS_MAX_VIEWER_NUM ];
 
-#ifdef __cplusplus
-}
-#endif
+    /* Media context. */
+    InitTransceiverFunc_t initTransceiverFunc;
+    AppMediaSourcesContext_t * pAppMediaSourcesContext;
+} AppContext_t;
 
-#endif /* DEMO_DATA_TYPES_H */
+int AppCommon_Init( AppContext_t * pAppContext,
+                    InitTransceiverFunc_t initTransceiverFunc,
+                    void * pMediaContext );
+int AppCommon_Start( AppContext_t * pAppContext );
+
+#endif /* APP_COMMON_H */
